third_party/blink/web_tests/wpt_internal/prerender/session-history-pushstate.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/session-history-test-util.js"></script>
<body>
  <script>
    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryPushStateInPrerender"),
        "Passed"
      );
    }, "history.pushState navigates independently with replacement in a prerender");

    promise_test(async () => {
      assert_equals(
        await runTestInPrerender("testHistoryReplaceStateInPrerender"),
        "Passed"
      );
    }, "history.replaceState navigates independently in a prerender");
  </script>
</body>
